requirements.txt
certifi==2018.8.24
cffi==1.11.5
cloudpickle==1.2.1
cycler==0.10.0
dask==2.1.0
decorator==4.4.0
imageio==2.3.0
kiwisolver==1.1.0
matplotlib==3.0.3
mkl-fft==1.0.9
mkl-random==1.0.1
networkx==2.3
numpy==1.15.2
olefile==0.46
Pillow==5.2.0
pycparser==2.19
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.1
scikit-image==0.14.0
scipy==1.1.0
six==1.11.0
toolz==0.10.0
torch==1.1.0
torchvision==0.3.0
tornado==5.1.1
